python/message-service/pyproject.toml
[tool.poetry]
name = "message-service"
version = "0.1.0"
description = ""
authors = ["Douglas Soo <doug@honeycomb.io>"]

[tool.poetry.dependencies]
python = "^3.8"
bottle = "^0.12.20"
opentelemetry-api = "^1.15.0"
opentelemetry-sdk = "^1.15.0"
opentelemetry-exporter-otlp-proto-grpc = "^1.15.0"
opentelemetry-instrumentation-requests = "^0.36b0"
opentelemetry-instrumentation-wsgi = "^0.36b0"
requests = "^2.28.1"
honeycomb-opentelemetry = {path = "../../../honeycomb-opentelemetry-python/dist/honeycomb_opentelemetry-0.1.0.tar.gz"}

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
